--- assets/.vscode/launch.json
@@ -1,16 +1,0 @@
-{
- // Use IntelliSense to learn about possible attributes.
- // Hover to view descriptions of existing attributes.
- // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
- "version": "0.2.0",
- "configurations": [
-  
-  {
-   "type": "chrome",
-   "request": "launch",
-   "name": "Launch Chrome against localhost",
-   "url": "http://localhost:8080",
-   "webRoot": "${workspaceFolder}"
-  }
- ]
-}
