--- assets/assets/scripts/style/css/style2.css
@@ -1,281 +1,0 @@
-body {
- background-color: #f1f1f1;
- font-family: Arial, Helvetica, sans-serif;
-}
-
-h1 {
- text-align: center;
- padding: 10px;
- background: #532855;
- color: #fff
-}
-
-table {
- width: 100%;
-}
-
-tr {
- text-align: center;
-}
-
-td {
- border: 5px solid transparent;
- background-color: #f1f1f1;
- padding: 10px;
- text-align: center;
- display: center;
- color: black;
- width: 25%;
- height: 40px;
-}
-
-td[class="grey"] {
- background-color: lightgrey;
-}
-
-td[class="blue"] {
- background-color: DodgerBlue;
-}
-
-td[class="orange"] {
- background-color: orange;
-}
-
-td[class="green"] {
- background-color: lightgreen;
-}
-
-div#turneu h2 {
- text-align: center;
- padding: 10px;
- background: #532855;
- color: #fff
-}
-
-div#turneu h3 {
- text-align: center;
- background: #283a55;
- color: #fff
-}
-
-.etapa {
- background: #283a55;
-}
-
-.match {
- box-shadow: rgba(0, 0, 0, 0.57) 0px 25px 20px -21px;
- border-radius: 10px;
- margin: 12px auto;
- background: #c8dfed;
-}
-
-.finala {
- box-shadow: rgba(0, 0, 0, 0.57) 0px 25px 20px -21px;
- border-radius: 10px;
- margin: 12px auto;
- background: #8ccdf5;
-}
-
-.match-info strong {
-
- color: #fff;
-
- position: absolute;
-
- right: 50%;
-
- text-align: center;
-
- z-index: 1;
-
- border-radius: 5px;
-
- height: 36px;
-
- line-height: 36px;
-
- top: 30px;
-
- width: 60px;
-
- margin-right: -30px
-}
-
-.match-info {
- position: relative;
- height: 160px;
- line-height: 50px
-}
-
-.right-team-info {
- float: right;
- padding-left: 30px;
- width: 50%;
- position: relative;
- z-index: 2
-}
-
-.left-team-info {
- float: left;
- padding-right: 0px;
- width: 50%;
- position: absolute;
- z-index: 2
-}
-
-.team-name {
- text-align: center;
- color: transparent;
- font-size: 20px;
- font-family: "Arial", sans-serif;
- font-weight: bold;
-}
-
-li {
- list-style-type: none;
- height: 20px;
- line-height: 20px;
- font-size: 14px;
- font-family: "Arial", sans-serif;
- font-weight: bold;
- text-align: center;
- padding-right: 30px;
-}
-
-.Verde {
- background-color: rgb(10, 145, 28);
- -webkit-background-clip: text;
-}
-
-.Portocaliu {
- background-color: rgb(207, 105, 8);
- -webkit-background-clip: text;
-}
-
-.Albastru {
- background-color: DodgerBlue;
- -webkit-background-clip: text;
-}
-
-.Gri {
- background-color: rgb(102, 90, 90);
- -webkit-background-clip: text;
-}
-
-.scor {
- background-color: #283a55;
-}
-
-.autocomplete-items {
- position: absolute;
- border: 1px solid #d4d4d4;
- border-bottom: none;
- border-top: none;
- z-index: 99;
- /*position the autocomplete items to be the same width as the container:*/
- top: 100%;
- left: 0;
- right: 0;
-}
-
-.autocomplete-items div {
- padding: 10px;
- cursor: pointer;
- background-color: #fff;
- border-bottom: 1px solid #d4d4d4;
-}
-
-/*when hovering an item:*/
-.autocomplete-items div:hover {
- background-color: #e9e9e9;
-}
-
-/*when navigating through the items using the arrow keys:*/
-.autocomplete-active {
- background-color: DodgerBlue !important;
- color: #ffffff;
-}
-
-.delete-button {
- font-size: 14px;
- color: red;
- margin-left: 5px;
- cursor: pointer;
-}
-
-
-input[type="text"] {
- padding: 5px;
- font-size: 16px;
- border: 1px solid #ccc;
- border-radius: 5px;
- width: 300px;
-}
-
-
-.recomandari {
- list-style: none;
- padding: 0;
- margin: 0;
- border: 1px solid #ccc;
- border-radius: 5px;
- background-color: #fff;
- max-height: 200px;
- overflow-y: auto;
- width: 298px;
- position: absolute;
- z-index: 6;
- opacity: 0;
-}
-
-.recomandari li {
- padding: 5px 10px;
- cursor: pointer;
-}
-
-.recomandari li:hover {
- background-color: #f9f9f9;
-}
-
-.recomandari:not(:empty) {
- opacity: 1;
- background-color: rgba(255, 255, 255, 1);
-}
-
-.recomandari:empty {
- opacity: 0;
- background-color: transparent;
-}
-
-button {
- padding: 8px 16px;
- font-size: 16px;
- border: none;
- border-radius: 5px;
- background-color: #007BFF;
- color: #fff;
- cursor: pointer;
-}
-
-button:hover {
- background-color: #0056b3;
-}
-
-.finala ul li {
- color: black;
-}
-
-.selectare {
- padding: 8px;
- font-size: 16px;
- border: 1px solid #ccc;
- border-radius: 5px;
- background-color: white;
- width: 100px;
-}
-
-
-.selectare option {
- padding: 8px;
-}
-
